Move MIPLIB metadata management into ommx::dataset::miplib
