+++ rust/ommx/Cargo.toml
@@ -20,6 +20,7 @@ base64.workspace = true
 chrono.workspace = true
 clap.workspace = true
 colored.workspace = true
+csv.workspace = true
 derive_more.workspace = true
 directories.workspace = true
 env_logger.workspace = true
